requirements.txt
numpy==1.26.4  #since meteostat requires numpy.NaN handling
pandas==2.2.2
requests==2.32.3
python-dateutil==2.9.0.post0
pytz==2024.1
meteostat==1.6.7
holidays==0.57
astral==3.2
ipykernel==6.29.5
statsmodels==0.14.2
matplotlib==3.9.2
seaborn==0.13.2
scikit-learn==1.4.1.post1
joblib==1.4.2